requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
PyMySQL
flask-login